add types ts to config directory
