--- federator/config/configChain.ts
@@ -1,10 +1,0 @@
-export class ConfigChain {
-  bridge: string;
-  nftBridge: string;
-  federation: string;
-  multiSig: string;
-  allowTokens: string;
-  testToken: string;
-  host: string;
-  fromBlock: number;
-}
