+++ federator/src/contracts/ContractFactory.ts
@@ -1,4 +1,4 @@
-import { Config } from '../../config/configts';
+import { Config } from '../../config/types';
 import Web3 from 'web3';
 import { AbiItem } from 'web3-utils';
 import { Contract } from 'web3-eth-contract';
